Add recentf function and save-places function into behaviour.
recentf - Allows quick access to files opened recently. Configuration
includes a limit on the number of files in the list (25) and placing
the recentf file in the ~/.Backups folder out of the way.

save-places - saves last location (specifically the last line number)
that the cursor was at when the file was closed. When the same file is
opened again the cursor is placed back on the same line. Handy for
very long files.