Added Palimpsest Functionality.
When I initially tried to install Palimpsest I got an error. Checking
the repository it appeared that Palimpsest wasn't in the melpa stable
repositories, and I needed to plumb in the normal package collection.

This involved changing the package archive information near the head
of the file.

Checking for available updates after moving to the unstable version of
the repository, it transpired 26 packages needed updating so this was
done.

I also added some instructions on using the Palimpsest plug-in.